Create GradebookChallengeProject
This is code I wrote for a challenge in the Khan Academy SQL course. The data table was created by a Khan Academy author and the code following the creation of the table is my own. I used the table to show different ways we could look at the grades of the students in the data table.